fix: address code review comments (PR #59, iteration 1)
- users.service: generate shareToken when inserting shareSettings with PRIVATE_LINK mode
- admin/users load: validate year param (2000-2100), fall back to current year on NaN/out-of-range
- admin/sync: make cronError a $derived so it stays in sync with presets and server updates
- auth/logout: convert +server.ts to +page.server.ts form action with redirect(303) for proper ActionResult + no-JS support
- onboarding/plex: post form-encoded body with x-sveltekit-action header to logout action
- rate-limit-handle: narrow /auth/plex special-case from startsWith to exact match so callback/redirect use stricter auth limit
