+++ src/lib/server/security/rate-limit-handle.ts
@@ -2,7 +2,7 @@ import type { Handle } from '@sveltejs/kit';
 import { checkRateLimit, RATE_LIMIT_CONFIGS, type RateLimitConfig } from '$lib/server/ratelimit';
 
 function getConfigForPath(path: string): RateLimitConfig {
-	if (path.startsWith('/auth/plex')) {
+	if (path === '/auth/plex') {
 		return RATE_LIMIT_CONFIGS.authPoll;
 	}
