+++ src/routes/admin/+layout.svelte
@@ -156,13 +156,7 @@ function handleCsrfWarningDismissed() {
 					<span class="user-role">Administrator</span>
 				</div>
 			</div>
-			<form method="POST" action="/auth/logout" use:enhance={() => {
-				return async ({ result }) => {
-					if (result.type === 'success') {
-						window.location.href = '/';
-					}
-				};
-			}}>
+			<form method="POST" action="/auth/logout" use:enhance>
 				<button type="submit" class="logout-button">
 					<LogOut class="logout-icon" />
 					<span>Logout</span>
